+++ amazon/dynamodb/client/src/test/kotlin/org/http4k/connect/amazon/dynamodb/ImportTableFromS3Contract.kt
@@ -166,17 +166,6 @@ private fun DynamoDb.waitForImportFinished(importArn: ARN, timeout: Duration = D
     )
 }
 
-private fun waitUntil(test: () -> Boolean, failureMessage: String, timeout: Duration) {
-    val waitStart = Instant.now()
-    while (Duration.between(waitStart, Instant.now()) < timeout) {
-        if (test()) {
-            return
-        }
-        Thread.sleep(1000)
-    }
-    throw IllegalStateException(failureMessage)
-}
-
 private fun DynamoDb.getItem(tableName: TableName, key: String, value: String) =
     getItem(tableName, Key(Attribute.string().required(key) of value)).successValue().item
